Finsih basic functionality

+++ src/app/calc-order-book/components/calc-order-book.component.ts
@@ -6,6 +6,7 @@ import { MolHttpClientErrorService } from 'src/app/shared/services/mol-http-clie
 import { CalcOrderBookService } from '../services/calc-order-book.service';
 import { CalcOrderViewModel } from '../view-models/calc-order-view-model';
 import { CalcOrderItemViewModel } from '../view-models/calc-order-item-view-model';
+import { UpdateOrderModalDlgComponent } from './update-order-modal-dlg/update-order-modal-dlg.component';
 
 @Component({
   selector: 'app-calc-order',
@@ -23,14 +24,11 @@ export class CalcOrderBookComponent implements OnInit{
                 private modalService:       NgbModal, 
                   private errorService:     MolHttpClientErrorService) { }
 
-  async ngOnInit() {
-      
+  async ngOnInit() {      
       this.errorService.SubscribeOnConnectionError((error) => this.displayError("No connection", error));
       this.errorService.SubscribeOnValidationError((error) => this.displayError("Vaidation error", error));
       this.errorService.SubscribeOnError((error) => this.displayError("Server error", error));
-
       await this.calcOrderBookService.GetCalculationOrders();
-
       this.SelectedOrder = this.Orders.length > 0 ? this.Orders[0] : null;
   }
 
@@ -45,6 +43,26 @@ export class CalcOrderBookComponent implements OnInit{
     );
   }
 
+  public OnClickEditOrder() {
+    if ( this.SelectedOrder) {
+      const modalRef = this.modalService.open(UpdateOrderModalDlgComponent);
+      modalRef.componentInstance.name = 'UpdateOrderModalDlgComponent';
+      modalRef.componentInstance.OrderName = this.SelectedOrder?.Details.Name;
+      modalRef.result.then( (result: CalcOrderViewModel) => {          
+        if ( this.SelectedOrder ) this.calcOrderBookService.UpdateCalculationOrder(this.SelectedOrder.Id, result.Details.Name, '')    
+      });
+    }
+  }
+
+  public OnClickDeleteOrder() {
+    if (  confirm("Are you sure you want to delete the order ?") && this.SelectedOrder) {
+       this.calcOrderBookService.DeleteCalculationOrder(this.SelectedOrder.Id)
+        .then(() => {
+          this.SelectedOrder = this.Orders.length > 0 ? this.Orders[0] : null;
+        })
+    }
+  }
+
   public OnClickNewOrderItem() {
     const modalRef = this.modalService.open(CreateOrderItemModalDlgComponent);
 		modalRef.componentInstance.name = 'CreateOrderItemModalDlgComponent';
@@ -57,7 +75,7 @@ export class CalcOrderBookComponent implements OnInit{
   }
 
   public OnClickDeleteOrderItem(item: CalcOrderItemViewModel) {
-    if( this.SelectedOrder) {
+    if( confirm("Are you sure you want to delete the order item ?") && this.SelectedOrder) {
       this.calcOrderBookService.DeleteOrderItem(this.SelectedOrder.Id, item.Id);
     }
   }
